Forms: Tests.iOS: Initialise using storyboard
