+++ tests/tests/Program.cs
@@ -26,25 +26,28 @@ using Microsoft.Xna.Framework.Content;
 namespace tests
 {
 #if IPHONE || IOS
+
     [Register ("AppDelegate")]
     internal class Program : UIApplicationDelegate 
     {
+        public override UIWindow Window {
+            get;
+            set;
+        }
+
+
+        public static UIStoryboard Storyboard = UIStoryboard.FromName ("TestsStoryboard", null);
+        public static UIViewController initialViewController;
+
+
         public override bool FinishedLaunching(UIApplication application, NSDictionary launchOptions)
         {
-            var window = new UIWindow (UIScreen.MainScreen.Bounds);
+            Window = new UIWindow (UIScreen.MainScreen.Bounds);
 
-            var vc = new UIViewController ();
+            initialViewController = Storyboard.InstantiateInitialViewController () as UIViewController;
 
-            window.RootViewController = vc;
-
-            CCGameView gameView = new CCGameView (window.Bounds);
-            gameView.ViewCreated += LoadGame;
-
-            vc.Add (gameView);
-
-            window.MakeKeyAndVisible ();
-
-            AppDelegate.SharedWindow = gameView;
+            Window.RootViewController = initialViewController;
+            Window.MakeKeyAndVisible ();
 
             return true;
         }
@@ -56,20 +59,6 @@ namespace tests
             // if you want to use a different Application Delegate class from "AppDelegate"
             // you can specify it here.
             UIApplication.Main (args, null, "AppDelegate");
-        }
-
-        void LoadGame(object sender, EventArgs e)
-        {
-            CCGameView gameView = sender as CCGameView;
-
-            if (gameView != null) 
-            {
-                gameView.DesignResolution = new CCSizeI (1024, 768);
-                gameView.Stats.Enabled = true;
-                CCScene gameScene = new CCScene (gameView);
-                gameScene.AddLayer(new TestController());
-                gameView.RunWithScene (gameScene);
-            }
         }
     }
 #endif
